+++ io-sim/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Added `ppSayTrace` which pritty prints `EventSay` which are coming from `say` usage.
 * Repository moved to https://github.com/IntersectMBO/io-sim
+* Exported `traceMarker`, `traceEvent` from `Control.Monad.Class.MonadEventlog`.
 
 ## 1.10.1.0